Adding updated Spin code
This version of the Spin source code now includes all the modules that
are needed, they just need to be "tied together". The demo application
still controls the 6502 via spin and basically shows what I had at the
2013 Bay Area Maker Faire Bring-a-hack dinner. But in theory there
should be enough there now to use the Hub module to map the demo code
and the video buffer, the Trace buffer to show what's going on, and the
Control module to single-step or run the demo code. The RAM module
should also be usable but nothing is tested yet. This is getting really
close to being able to running at full speed!