Updating the C code
The C code depends on the new .pasm directive in the GCC compiler and it
looks like it compiles correctly, but because the .pasm directive is
experimental in the Propeller GCC compiler, there's a good chance that
code gets compiled incorrectly even though it's logically correct. For
now I don't want to deal with the possible debugging issues that might
be involved with that, so I'll just check this in by way of archiving
it. The most up-to-date software will be in the software/spin directory
from this point on. Once I get the spin/pasm based code to work
reliably, I will translate it to C/pasm again, for the purpose of
generating code that's more flexible (because C has a preprocessor that
does a better job of keeping things in sync between modules, than
trusting my memory) and can be used to verify the correct operation of
the GCC compiler, which I totally support.