More updates to schematic
I decided to keep working on the schematic and PCB layout in Kicad until
I have the current schematic (Rev. 8A which needed a Propeller Platform
motherboard and a modified VGA/AV module) implemented as Kicad files,
and it completely matches what I have in Eagle. After that, I'll take
the next step which is to design a new PCB with all the necessary
hardware. That way, I can learn Kicad and focus on problems caused by
the imperfect export script that I used to convert the Eagle design.
Debugging things like short-circuits caused by Kicad interpreting two
bus connections in the same place in a different way than Eagle, is much
easier when the schematic is small and familiar to myself. Another
problem was that the Eagle export script had implemented the power pins
of the IC's as multiple units within those IC's on Kicad. But where the
IC's power connections were in the schematic, it drew the entire chip
instead of just the power pins (it selected the wrong unit apparently).
I had deleted those messy parts in the previous check-in but I tried
today to recover the power connections. First I tried to implement them
as invisible pins, as is apparently the standard in Kicad (and I'm not
sure if I like that yet) but my IC power pins weren't being connected to
the same net as the VCC and GND pins elsewhere in the schematic
according to the rats nest in PcbNew. I eventually made all the power
pins visible and hooked them up in the schematic again (very ugly) and
it still wasn't working. After I made a test project and inspected the
differences between the .net files, I got the idea of deleting the VCC
and GND *symbols* from the propeddle-cache.lib because apparently they
were interfering with how the power.lib symbols were working. After I
did this and saved the library, and changed the schematic so that the
power.lib symbols were properly connected (they're smaller so I needed
to wire them all up), all weird DRC errors disappeared, and I could add
the PWR_FLAG symbols that all tutorials are talking about, and PcbNew
finally showed the power pins as connected. Yay! Time to commit that.