Completed design of "all-in-one" schematic/PCB
Some of the important changes:
- Everything on one PCB including Propeller, connectors and power
supply. The only thing missing is the serial-to-USB converter.
- Added inverters: the signals (~IRQ, ~NMI, ~RES, ~SO and RDY) are now
connected via open-drain inverters which eliminates all diodes in the
old schematic.
- The old SEL0 and SEL1 are now replaced by BE (via open-drain inverter)
and SETUP. These signals will make it possible to initialize expansion
hardware via the data bus.
- Added power supply, Propeller, connectors for keyboard, mouse, TV and
power (and added their documentation)
- The expansion port was redesigned (Sorry, future expansion hardware
will not be compatible with the previous Propeddle boards)