Fixed timing bug in Trace cog
I forgot that when copying the system counter to a variable, CNT is
copied at the SECOND cycle, not the first cycle. This threw the timing
off by one cycle, which made it pick up the data bus value one cycle too
late if the system is running at 80 cycles; also when that happens it
misses a deadline to pick up the next cycle's tracing information and
everything goes haywire.