Change frequency to 100MHz, Add terminal module
I realized that now the Propeller circuit is part of the kit, I'm not
limited to 80MHz anymore so I'm free to change the BOM so it uses a
6.25MHz crystal instead of a 5MHz crystal which may give me a little
more leeway in cases where timing is tight. From now on, the code will
be tested with a 6.25MHz system (internal frequency 100MHz). The
comments and code aren't completely adapted to the new standard yet.

A new module was added that emulates the keyboard and display hardware
of the Apple 1. The keyboard part hasn't been tested yet, but the test
program in this commit demonstrates the output functionality: it
repeatedly prints all printable characters to the terminal. The 6502
source code is roughly as follows:

RESET:
LDA #$20
LOOP
TAX
CMP #$7F
BEQ RESET
JSR ECHO
INX
TXA
JMP LOOP
ECHO
BIT $D012
BMI ECHO
STA $D012
RTS

Note: the ECHO routine is identical to the routine with the same name in
the Woz monitor ROM on the Apple 1 and is stored in the same location:
$FFEF.