Re-adding AIGeneric driver
The AIGeneric driver was downloaded from the Parallax Object Exchange
(OBEX) at http://obex.parallax.com/object/139.
My SerInTVOut driver uses it to allow input from the serial port and the
PS/2 keyboard at the same time and output to the serial port and screen
at the same time. At this time, I'm away from home so I can't test the
PS/2 and video part; the serial communications work (115200,8,n,1).
There is a know issue that when the serial port is not connected, the
code thinks there are characters coming in continuously.